# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/create-user/{admin_mail}:
    post:
      operationId: AuthController_createCompanyUser
      parameters:
        - name: admin_mail
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Details for creating a company user
        content:
          application/json:
            schema:
              type: object
              properties:
                company_ifric_id:
                  type: string
                  example: IFRIC12345
                user_name:
                  type: string
                  example: John Doe
                user_email:
                  type: string
                  example: johndoe@example.com
                user_password:
                  type: string
                  example: securepassword123
                products:
                  type: array
                  items:
                    type: object
                    properties:
                      product_name:
                        type: string
                        example: ProductA
                      product_version:
                        type: string
                        example: v1.0
                  example:
                    - product_name: ProductA
                      product_version: v1.0
                    - product_name: ProductB
                      product_version: v2.1
              required:
                - company_ifric_id
                - user_name
                - user_email
                - user_password
                - products
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/login:
    post:
      operationId: AuthController_logIn
      parameters: []
      requestBody:
        required: true
        description: User login details
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: securepassword123
                product_name:
                  type: string
                  example: Product A
              required:
                - email
                - password
                - product_name
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-indexed-db-data:
    post:
      operationId: AuthController_getIndexedData
      parameters: []
      requestBody:
        required: true
        description: Retrieve indexed database data
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                  description: The email of the user
                product_name:
                  type: string
                  example: Product A
                  description: The name of the product
                company_id:
                  type: string
                  example: COMPANY_ID_123
                  description: The unique ID of the company
              required:
                - email
                - product_name
                - company_id
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/authenticate-token/{token}:
    get:
      operationId: AuthController_authenticateToken
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-company-users/{id}:
    get:
      operationId: AuthController_getCompanyUsers
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-company-users-access/{company_ifric_id}:
    get:
      operationId: AuthController_getCompanyUsersAccess
      parameters:
        - name: company_ifric_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-user-profile-content/{company_ifric_id}/{user_id}:
    get:
      operationId: AuthController_getUserProfileContent
      parameters:
        - name: company_ifric_id
          required: true
          in: path
          schema:
            type: string
        - name: user_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-user-product-access/{id}:
    get:
      operationId: AuthController_getUserProductAccess
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-user-details:
    get:
      operationId: AuthController_getUserDetails
      parameters:
        - name: user_email
          required: true
          in: query
          schema:
            type: string
        - name: company_ifric_id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-total-users:
    get:
      operationId: AuthController_getTotalUsers
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-user-details/{id}:
    get:
      operationId: AuthController_getUserDetailsById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-user-details-by-email/{email}:
    get:
      operationId: AuthController_getUserDetailsByEmail
      parameters:
        - name: email
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-user-details-by-email-recover-password/{email}:
    get:
      operationId: AuthController_getUserDetailsForRecoverPassword
      parameters:
        - name: email
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/get-user-specific-product-access:
    get:
      operationId: AuthController_getUserSpecificProductAccess
      parameters:
        - name: product_name
          required: true
          in: query
          schema:
            type: string
        - name: user_id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/check-company-admin/{email}:
    get:
      operationId: AuthController_checkCompanyAdmin
      parameters:
        - name: email
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/update-password:
    patch:
      operationId: AuthController_updateUserPassword
      parameters: []
      requestBody:
        required: true
        description: Change a Company or CompanyUser password, verifying the old password first
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                  description: The email of the user
                oldPassword:
                  type: string
                  example: OldWeakP@ssw0rd!
                  description: The current password, verified before updating
                newPassword:
                  type: string
                  example: NewStrongP@ssw0rd!
                  description: The new password for the user
              required:
                - email
                - oldPassword
                - newPassword
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/update-user-access-group/{id}:
    patch:
      operationId: AuthController_updateUserAccessGroup
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: For each entry, grants the user (CompanyUser :id) the named access group role on the named product
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  product:
                    type: string
                    example: Software
                    description: Product name
                  user_role:
                    type: string
                    example: admin
                    description: Access group name (matches AccessGroup.group_name)
                required:
                  - product
                  - user_role
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/update-company-user:
    patch:
      operationId: AuthController_updateCompanyUser
      parameters: []
      requestBody:
        required: true
        description: Update user details with optional fields for user information and password changes
        content:
          application/json:
            schema:
              type: object
              properties:
                company_ifric_id:
                  type: string
                  example: company-ifric-123
                  description: Unique identifier for the company
                user_id:
                  type: string
                  example: user-456
                  description: Unique identifier for the user
                user_name:
                  type: string
                  example: John Doe
                  description: Name of the user
                user_email:
                  type: string
                  example: john.doe@example.com
                  description: Email address of the user
                user_image:
                  type: string
                  example: https://example.com/path/to/image.jpg
                  description: URL to the user's profile image
                old_password:
                  type: string
                  example: OldPassword123
                  description: Old password for the user (if changing password)
                new_password:
                  type: string
                  example: NewPassword456
                  description: New password for the user (if changing password)
                jwt_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR...
                  description: JWT token for authenticating the user
              required:
                - company_ifric_id
                - user_id
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/delete-company-user/{id}:
    delete:
      operationId: AuthController_deleteCompanyUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/logout:
    post:
      operationId: AuthController_logOut
      parameters: []
      requestBody:
        required: true
        description: Log out a user by providing their email address
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                  description: The email address of the user to log out
              required:
                - email
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        description: Exchange a refresh token (obtained from /auth/login) for a new, short-lived access token. Fails if the refresh token has expired or been revoked (e.g. by /auth/logout).
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: The refresh token obtained from /auth/login
              required:
                - refresh_token
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/recover-password-request:
    post:
      operationId: AuthController_recoverPasswordRequest
      parameters: []
      requestBody:
        required: true
        description: Generate a new temporary password for a user, returned directly in the response (no email is sent — this project has no bundled email concept). The caller is responsible for relaying it to the user.
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                  description: The email address of the user requesting password recovery
              required:
                - email
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
  /auth/recover-password:
    post:
      operationId: AuthController_recoverPassword
      parameters: []
      requestBody:
        required: true
        description: Recover a user password using a temporary password and set a new password
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                  description: The email address of the user
                temporaryPassword:
                  type: string
                  example: Temp@1234
                  description: The temporary password sent to the user
                newPassword:
                  type: string
                  example: NewStrongP@ssw0rd!
                  description: The new password to be set for the user
              required:
                - email
                - temporaryPassword
                - newPassword
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Auth
...
info:
  title: Ifric Registry Service
  description: Open-source multi-tenant company/user/access-control/digital-twin registry service, split into Auth (session/identity), Company (company CRUD, access groups, physical assets), Product (product catalog, digital twins), Certificate, and Script (one-time seed data) controllers. All use a single bearer-JWT auth mechanism (POST /auth/login, refreshed via POST /auth/refresh). Certificate issuance and company_ifric_id minting integrate with an external open-source ICID-compatible service (see README.md).
  version: 0.1.0
  contact: {}
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
...